requirements.txt
Flask
requests
python-dotenv
Werkzeug
gunicorn
google-generativeai